Factor export removal out of action
