+++ client/src/components/BlockEditor/types.tsx
@@ -17,7 +17,9 @@ export interface Change {
   moveFocus(n: number): Change;
   moveFocusToEndOfPreviousText(): Change;
   moveToRangeOf(node: any): Change;
+  removeNodeByKey(key: string): Change;
   removeTextByKey(key: string, offset: number, length: number): Change;
+  unwrapInlineByKey(key: string, properties: any): Change;
 }
 
 export interface TextNode {
