+++ client/src/modules/blockEditor/actions.ts
@@ -67,7 +67,7 @@ export const exportSelection = () => {
       const change = block.value.change();
       slateChangeMutations.insertPointerExport(change);
       slateChangeMutations.normalizeExportSpacing(change);
-      
+
       dispatch({
         type: UPDATE_BLOCK,
         id: block.id,
@@ -80,24 +80,6 @@ export const exportSelection = () => {
   };
 };
 
-function getInlinesAsArray(node: any) {
-  let array: any = [];
-
-  node.nodes.forEach(child => {
-    if (child.object === "text") {
-      return;
-    }
-    if (child.object === "inline") {
-      array.push(child);
-    }
-    if (!child.isLeafInline()) {
-      array = array.concat(getInlinesAsArray(child));
-    }
-  });
-
-  return array;
-}
-
 export const removeExportOfSelection = () => {
   return async (dispatch, getState) => {
     const { blocks, blockEditor } = await getState();
@@ -106,32 +88,8 @@ export const removeExportOfSelection = () => {
     const block = blocks.blocks.find(b => b.id === hoveredItem.blockId);
 
     if (block) {
-      const inlines = getInlinesAsArray(block.value.document);
-      const matchingNodes = inlines.filter(
-        i => i.toJSON().data.pointerId === hoveredItem.id
-      );
-      if (!matchingNodes.length) {
-        console.error("Exporting node not found in Redux store");
-        return;
-      }
-
-      const nodeToRemove = matchingNodes[0];
-
-      let change;
-      change = block.value.change();
-
-      // remove pointerExport inline
-      const ancestorNodes = block.value.document.getAncestors(nodeToRemove.key);
-      const isNested = ancestorNodes.find(ancestor => ancestor.type === "pointerExport");
-
-      if (isNested) {
-        change
-          .unwrapInlineByKey(nodeToRemove.key,  { data: { pointerId: hoveredItem.id } })
-          .removeNodeByKey(nodeToRemove.key);
-      } else {
-        change.unwrapInlineByKey(nodeToRemove.key,  { data: { pointerId: hoveredItem.id } });
-      }
-
+      const change = block.value.change();
+      slateChangeMutations.removePointerExport(change, hoveredItem);
       slateChangeMutations.normalizeAfterRemoval(change);
 
       dispatch({
